# [Ready] Reduce reload word tokenizer engine in word_tokenize