### What does this changes

Fixes #973 

### What was wrong

<!--- Description of what was the root cause of the issue. -->

### How this fixes it

<!--- Description of how the changes fix the issue. -->

Fixes #...

### Your checklist for this pull request

<!--- Please review the guidelines for contributing to this repository at: -->
<!--- https://github.com/PyThaiNLP/pythainlp/blob/dev/CONTRIBUTING.md -->

- [x] Passed code styles and structures
- [x] Passed code linting checks and unit test
- [x] Test Attacut and icu manually
- [x] Benchmark speed improvement