@pep8speaks 2025-01-11
Hello @new5558! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`pythainlp/tokenize/attacut.py`](https://github.com/PyThaiNLP/pythainlp/blob/9efd6e77413f0d32aa270af9c7021307607b3618/pythainlp/tokenize/attacut.py):

> [Line 28:1](https://github.com/PyThaiNLP/pythainlp/blob/9efd6e77413f0d32aa270af9c7021307607b3618/pythainlp/tokenize/attacut.py#L28): [E305](https://duckduckgo.com/?q=pep8%20E305) expected 2 blank lines after class or function definition, found 1
> [Line 44:1](https://github.com/PyThaiNLP/pythainlp/blob/9efd6e77413f0d32aa270af9c7021307607b3618/pythainlp/tokenize/attacut.py#L44): [W293](https://duckduckgo.com/?q=pep8%20W293) blank line contains whitespace





* In the file [`tests/core/test_tokenize.py`](https://github.com/PyThaiNLP/pythainlp/blob/9efd6e77413f0d32aa270af9c7021307607b3618/tests/core/test_tokenize.py):

> [Line 419:14](https://github.com/PyThaiNLP/pythainlp/blob/9efd6e77413f0d32aa270af9c7021307607b3618/tests/core/test_tokenize.py#L419): [E131](https://duckduckgo.com/?q=pep8%20E131) continuation line unaligned for hanging indent
> [Line 423:14](https://github.com/PyThaiNLP/pythainlp/blob/9efd6e77413f0d32aa270af9c7021307607b3618/tests/core/test_tokenize.py#L423): [E131](https://duckduckgo.com/?q=pep8%20E131) continuation line unaligned for hanging indent