@new5558 2025-01-11
Manual test of Attacut & ICU tokenizers results and benchmark results can be found in this [Colab Notebook](https://colab.research.google.com/drive/1hywE445wldiUKHpeHB9kSF1I7nmtWoT8?usp=sharing)

### Performance Improvements

| Tokenizer | Original (ms) | This PR (ms) | Original 50 threads (ms) | This PR 50 threads (ms) |
|-----------|---------------|--------------|---------------------------|--------------------------|
| Attacut   | 9.83 ms       | **1.67 ms**  | 17.18 ms                  | **2.83 ms**             |
| ICU       | 0.03 ms       | 0.03 ms      | 0.19 ms                   | 0.27 ms                 |
| Longest   | 0.02 ms       | 0.01 ms  | 0.21 ms                   | 0.20 ms             |

Attacut engine is 6X Faster when tokenizing words `"สวัสดี PythaiNLP"`.
Not much improvement in ICU and Longest engines,